sbctl-batch-sign: Skip signing files in /boot/grub
GRUB has EFI binaries in this path, but they're not actually used[1] for booting so we don't need to sign them.

[1] Past users who used secure boot with GRUB only signed the EFI binaries in /boot/efi and kernel images.
I was only aware of these files when I tested these changes in my VM. Unfortunately, I couldn't get secure boot
to work on that VM, but the script should work fine.

Signed-off-by: Eric Naim <dnaim@cachyos.org>